Removes all props from CharacterMovesForm
